packages/mas-server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mas-server"
version = "0.4.5"
description = "MAS server runtime"
requires-python = ">=3.13"
dependencies = [
    "grpcio>=1.76.0",
    "mas-core",
    "mas-gateway",
    "mas-proto",
]

[tool.uv.sources]
mas-core = { workspace = true }
mas-gateway = { workspace = true }
mas-proto = { workspace = true }

[tool.hatch.build.targets.wheel]
packages = ["src/mas_server"]
